+++ unit-tests/CMakeLists.txt
@@ -15,8 +15,10 @@ endif()
 
 if(BUILD_UNIT_TESTS)
     if(Python3_FOUND)
+		if( NOT DEFINED UNIT-TESTS )
+			set( UNIT-TESTS "" )
         execute_process(
-            COMMAND ${Python3_EXECUTABLE} ${CMAKE_CURRENT_SOURCE_DIR}/unit-test-config.py ${CMAKE_CURRENT_SOURCE_DIR} ${CMAKE_CURRENT_BINARY_DIR}
+            COMMAND ${Python3_EXECUTABLE} ${CMAKE_CURRENT_SOURCE_DIR}/unit-test-config.py ${CMAKE_CURRENT_SOURCE_DIR} ${CMAKE_CURRENT_BINARY_DIR} ${UNIT-TESTS}
             RESULT_VARIABLE rv
         )
         if( NOT ${rv} EQUAL 0 )
