+++ unit-tests/run-unit-tests.py
@@ -4,7 +4,6 @@
 # Copyright(c) 2021 Intel Corporation. All Rights Reserved.
 
 import sys, os, subprocess, locale, re, platform, getopt, time
-from abc import ABC, abstractmethod
 
 # Remove Python's default list of places to look for modules!
 # We want only modules in the directories we specifically provide to be found,
@@ -18,7 +17,7 @@ sys.path.append( os.path.join( os.path.dirname( sys.executable ), 'lib' ) )
 current_dir = os.path.dirname( os.path.abspath( __file__ ) )
 sys.path.append( current_dir + os.sep + "py" )
 
-from rspy import log, file, repo
+from rspy import log, file, repo, unittest
 
 
 def usage():
@@ -57,7 +56,6 @@ except getopt.GetoptError as err:
     log.e( err )  # something like "option -a not recognized"
     usage()
 regex = None
-to_stdout = False
 required_tags = []
 list_tags = False
 list_tests = False
@@ -71,7 +69,7 @@ for opt, arg in opts:
     elif opt in ('-r', '--regex'):
         regex = arg
     elif opt in ('-s', '--stdout'):
-        to_stdout = True
+        unittest.to_stdout = True
     elif opt in ('-t', '--tag'):
         required_tags.append( arg )
     elif opt == '--list-tags':
@@ -105,6 +103,7 @@ if target:
 else:  # no test executables were found. We put the logs directly in build directory
     logdir = os.path.join( repo.root, 'build', 'unit-tests' )
 os.makedirs( logdir, exist_ok=True )
+unittest.logdir = logdir
 n_tests = 0
 
 # Python scripts should be able to find the pyrealsense2 .pyd or else they won't work. We don't know
@@ -138,52 +137,6 @@ os.environ["PYTHONPATH"] = current_dir + os.sep + "py"
 #
 if pyrs:
     os.environ["PYTHONPATH"] += os.pathsep + pyrs_path
-
-
-def subprocess_run( cmd, stdout = None, timeout = 200, append = False ):
-    """
-    Wrapper function for subprocess.run.
-    If the child process times out or ends with a non-zero exit status an exception is raised!
-
-    :param cmd: the command and argument for the child process, as a list
-    :param stdout: path of file to direct the output of the process to (None to disable)
-    :param timeout: number of seconds to give the process before forcefully ending it (None to disable)
-    :param append: if True and stdout is not None, the log of the test will be appended to the file instead of
-                   overwriting it
-    :return: the output written by the child, if stdout is None -- otherwise N/A
-    """
-    log.d( 'running:', cmd )
-    handle = None
-    start_time = time.time()
-    try:
-        log.debug_indent()
-        if stdout and stdout != subprocess.PIPE:
-            if append:
-                handle = open( stdout, "a" )
-                handle.write(
-                    "\n---------------------------------------------------------------------------------\n\n" )
-                handle.flush()
-            else:
-                handle = open( stdout, "w" )
-            stdout = handle
-        rv = subprocess.run( cmd,
-                             stdout=stdout,
-                             stderr=subprocess.STDOUT,
-                             universal_newlines=True,
-                             timeout=timeout,
-                             check=True )
-        result = rv.stdout
-        if not result:
-            result = []
-        else:
-            result = result.split( '\n' )
-        return result
-    finally:
-        if handle:
-            handle.close()
-        log.debug_unindent()
-        run_time = time.time() - start_time
-        log.d( "test took", run_time, "seconds" )
 
 
 def configuration_str( configuration, prefix = '', suffix = '' ):
@@ -238,294 +191,6 @@ def check_log_for_fails( path_to_log, testname, configuration = None ):
     return False
 
 
-class TestConfig( ABC ):  # Abstract Base Class
-    """
-    Configuration for a test, encompassing any metadata needed to control its run, like retries etc.
-    """
-
-    def __init__( self ):
-        self._configurations = list()
-        self._priority = 1000
-        self._tags = set()
-        self._flags = set()
-        self._timeout = 200
-
-    def debug_dump( self ):
-        if self._priority != 1000:
-            log.d( 'priority:', self._priority )
-        if self._timeout != 200:
-            log.d( 'timeout:', self._timeout )
-        if len( self._tags ) > 1:
-            log.d( 'tags:', { tag for tag in self._tags if tag != "exe" and tag != "py" } )
-        if self._flags:
-            log.d( 'flags:', self._flags )
-        if len( self._configurations ) > 1:
-            log.d( len( self._configurations ), 'configurations' )
-            # don't show them... they are output separately
-
-    @property
-    def configurations( self ):
-        return self._configurations
-
-    @property
-    def priority( self ):
-        return self._priority
-
-    @property
-    def timeout( self ):
-        return self._timeout
-
-    @property
-    def tags( self ):
-        return self._tags
-
-    @property
-    def flags( self ):
-        return self._flags
-
-
-class TestConfigFromText( TestConfig ):
-    """
-    Configuration for a test -- from any text-based syntax with a given prefix, e.g. for python:
-        #test:usb2
-        #test:device L500* D400*
-        #test:retries 3
-        #test:priority 0
-    And, for C++ the prefix could be:
-        //#test:...
-    """
-
-    def __init__( self, source, line_prefix ):
-        """
-        :param source: The path to the text file
-        :param line_prefix: A regex to denote a directive (must be first thing in a line), which will
-            be immediately followed by the directive itself and optional arguments
-        """
-        TestConfig.__init__( self )
-
-        # Parse the python
-        regex = r'^' + line_prefix + r'(\S+)((?:\s+\S+)*?)\s*(?:#\s*(.*))?$'
-        for context in file.grep( regex, source ):
-            match = context['match']
-            directive = match.group( 1 )
-            text_params = match.group( 2 ).strip()
-            params = [s for s in text_params.split()]
-            comment = match.group( 3 )
-            if directive == 'device':
-                # log.d( '    configuration:', params )
-                if not params:
-                    log.e( source + '+' + str( context['index'] ) + ': device directive with no devices listed' )
-                elif 'each' in text_params.lower() and len( params ) > 1:
-                    log.e( source + '+' + str(
-                        context['index'] ) + ': each() cannot be used in combination with other specs', params )
-                elif 'each' in text_params.lower() and not re.fullmatch( r'each\(.+\)', text_params, re.IGNORECASE ):
-                    log.e( source + '+' + str( context['index'] ) + ': invalid \'each\' syntax:', params )
-                else:
-                    self._configurations.append( params )
-            elif directive == 'priority':
-                if len( params ) == 1 and params[0].isdigit():
-                    self._priority = int( params[0] )
-                else:
-                    log.e( source + '+' + str( context['index'] ) + ': priority directive with invalid parameters:',
-                           params )
-            elif directive == 'timeout':
-                if len( params ) == 1 and params[0].isdigit():
-                    self._timeout = int( params[0] )
-                else:
-                    log.e( source + '+' + str( context['index'] ) + ': timeout directive with invalid parameters:',
-                           params )
-            elif directive == 'tag':
-                self._tags.update( params )
-            elif directive == 'flag':
-                self._flags.update( params )
-            else:
-                log.e( source + '+' + str( context['index'] ) + ': invalid directive "' + directive + '"; ignoring' )
-
-
-class Test( ABC ):  # Abstract Base Class
-    """
-    Abstract class for a test. Holds the name of the test
-    """
-
-    def __init__( self, testname ):
-        # log.d( 'found', testname )
-        self._name = testname
-        self._config = None
-        self._ran = False
-
-    @abstractmethod
-    def run_test( self, configuration = None, log_path = None ):
-        pass
-
-    def debug_dump( self ):
-        if self._config:
-            self._config.debug_dump()
-
-    @property
-    def config( self ):
-        return self._config
-
-    @property
-    def name( self ):
-        return self._name
-
-    @property
-    def ran( self ):
-        return self._ran
-
-    def get_log( self ):
-        global to_stdout
-        if to_stdout:
-            path = None
-        else:
-            path = logdir + os.sep + self.name + ".log"
-        return path
-
-    def is_live( self ):
-        """
-        Returns True if the test configurations specify devices (test has a 'device' directive)
-        """
-        return self._config and len( self._config.configurations ) > 0
-
-    def find_source_path( self ):
-        """
-        :return: The relative path from unit-tests directory to the test's source file (cpp or py). If the source
-                 file is not found None will be returned
-        """
-        # TODO: this is limited to a structure in which .cpp files and directories do not share names
-        # For example:
-        #     unit-tests/
-        #         func/
-        #             ...
-        #         test-func.cpp
-        # test-func.cpp will not be found!
-
-        global current_dir
-        split_testname = self.name.split( '-' )
-        path = current_dir
-        relative_path = ""
-        found_test_dir = False
-
-        while not found_test_dir:
-            # index 0 should be 'test' as tests always start with it
-            found_test_dir = True
-            for i in range( 2,
-                            len( split_testname ) ):  # Checking if the next part of the test name is a sub-directory
-                possible_sub_dir = '-'.join( split_testname[1:i] )  # The next sub-directory could have several words
-                sub_dir_path = path + os.sep + possible_sub_dir
-                if os.path.isdir( sub_dir_path ):
-                    path = sub_dir_path
-                    relative_path += possible_sub_dir + os.sep
-                    del split_testname[1:i]
-                    found_test_dir = False
-                    break
-
-        path += os.sep + '-'.join( split_testname )
-        relative_path += '-'.join( split_testname )
-        if os.path.isfile( path + ".cpp" ):
-            relative_path += ".cpp"
-        elif os.path.isfile( path + ".py" ):
-            relative_path += ".py"
-        else:
-            log.w( log.red + self.name + log.reset + ':',
-                   'No matching .cpp or .py file was found; no configuration will be used!' )
-            return None
-
-        return relative_path
-
-
-class PyTest( Test ):
-    """
-    Class for python tests. Hold the path to the script of the test
-    """
-
-    def __init__( self, testname, path_to_test ):
-        """
-        :param testname: name of the test
-        :param path_to_test: the relative path from the current directory to the path
-        """
-        global current_dir
-        Test.__init__( self, testname )
-        self.path_to_script = current_dir + os.sep + path_to_test
-        self._config = TestConfigFromText( self.path_to_script, r'#\s*test:' )
-        relative_test_path = self.find_source_path()
-        if relative_test_path:
-            sub_dirs = relative_test_path.split( os.sep )[:-1] # last element will be the test's name
-            self._config.tags.update( sub_dirs )
-        self._config.tags.add( "py" )
-
-    def debug_dump( self ):
-        log.d( 'script:', self.path_to_script )
-        Test.debug_dump( self )
-
-    @property
-    def command( self ):
-        cmd = [sys.executable]
-        # The unit-tests should only find module we've specifically added -- but Python may have site packages
-        # that are automatically made available. We want to avoid those:
-        #     -S     : don't imply 'import site' on initialization
-        # NOTE: exit() is defined in site.py and works only if the site module is imported!
-        cmd += ['-S']
-        if sys.flags.verbose:
-            cmd += ["-v"]
-        cmd += [self.path_to_script]
-        if 'custom-args' not in self.config.flags:
-            if log.is_debug_on():
-                cmd += ['--debug']
-            if log.is_color_on():
-                cmd += ['--color']
-        return cmd
-
-    def run_test( self, configuration = None, log_path = None ):
-        try:
-            subprocess_run( self.command, stdout=log_path, append=self.ran, timeout=self.config.timeout )
-        finally:
-            self._ran = True
-
-
-class ExeTest( Test ):
-    """
-    Class for c/cpp tests. Hold the path to the executable for the test
-    """
-
-    def __init__( self, testname, exe ):
-        """
-        :param testname: name of the test
-        :param exe: full path to executable
-        """
-        global current_dir
-        Test.__init__( self, testname )
-        self.exe = exe
-
-        relative_test_path = self.find_source_path()
-        if relative_test_path:
-            sub_dirs = relative_test_path.split( os.sep )[:-1]  # last element will be the test's name
-            self._config = TestConfigFromText( current_dir + os.sep + relative_test_path, r'//#\s*test:' )
-            self._config.tags.update( sub_dirs )
-            self._config.tags.add( "exe" )
-
-
-    @property
-    def command( self ):
-        cmd = [self.exe]
-        if 'custom-args' not in self.config.flags:
-            # Assume we're a Catch2 exe, so:
-            # if sys.flags.verbose:
-            #    cmd += 
-            if log.is_debug_on():
-                cmd += ['-d', 'yes']  # show durations for each test-case
-                # cmd += ['--success']  # show successful assertions in output
-            # if log.is_color_on():
-            #    cmd += ['--use-colour', 'yes']
-        return cmd
-
-    def run_test( self, configuration = None, log_path = None ):
-        try:
-            subprocess_run( self.command, stdout=log_path, append=self.ran, timeout=self.config.timeout )
-        finally:
-            self._ran = True
-
-
 def get_tests():
     global regex, target, pyrs, current_dir, linux
     if regex:
@@ -558,7 +223,7 @@ def get_tests():
             else:
                 exe = target + '/' + testname + '.exe'
 
-            yield ExeTest( testname, exe )
+            yield unittest.ExeTest( testname, exe )
 
     # Python unit-test scripts are in the same directory as us... we want to consider running them
     # (we may not if they're live and we have no pyrealsense2.pyd):
@@ -572,7 +237,7 @@ def get_tests():
         if regex and not pattern.search( testname ):
             continue
 
-        yield PyTest( testname, py_test )
+        yield unittest.PyTest( testname, py_test )
 
 
 def prioritize_tests( tests ):
