+++ unit-tests/unit-test-config.py
@@ -94,7 +94,7 @@ src = librealsense + '/src'
 def generate_cmake( builddir, testdir, testname, filelist ):
     makefile = builddir + '/' + testdir + '/CMakeLists.txt'
     debug( '   creating:', makefile )
-    handle = open( makefile, 'w' );
+    handle = open( makefile, 'w' )
     filelist = '\n    '.join( filelist )
     handle.write( '''
 # This file is automatically generated!!
@@ -146,9 +146,14 @@ def process_cpp( dir, builddir ):
         testdir = os.path.splitext( f )[0]                          # "log/internal/test-all"  <-  "log/internal/test-all.cpp"
         testparent = os.path.dirname(testdir)                       # "log/internal"
         # Each CMakeLists.txt sits in its own directory
-        os.makedirs( builddir + '/' + testdir, exist_ok = True );   # "build/log/internal/test-all"
+        os.makedirs( builddir + '/' + testdir, exist_ok = True )    # "build/log/internal/test-all"
         # We need the project name unique: keep the path but make it nicer:
         testname = 'test-' + testparent.replace( '/', '-' ) + '-' + os.path.basename(testdir)[5:]   # "test-log-internal-all"
+
+        #TODO: 1. check if testname matches the regex if -r was used
+        #      2. if -t was used, create ExeTest using the testname and check if the tags fit
+        #    if either of these don't match, continue
+
         # Build the list of files we want in the project:
         # At a minimum, we have the original file, plus any common files
         filelist = [ dir + '/' + f, '${ELPP_FILES}', '${CATCH_FILES}' ]
